+++ jest.config.js
@@ -19,7 +19,8 @@ module.exports = {
   // But its performance overhead is pretty bad (30+%).
   // detectOpenHandles: true
 
-  setupFilesAfterEnv: ['./test/utils/setup'],
+  setupFilesAfterEnv: ['./test/utils/setup-env'],
+  setupFiles: ['./test/utils/setup'],
 
   coverageDirectory: './coverage',
